# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Orchestration backend for the automated short-form video pipeline, plus the original book sample endpoints.",
        "title": "Pom Pom Hollow Production API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/pipeline/episodes": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "List every episode, newest first. Filter by pipeline stage with ?status=",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pipeline-episodes"
                ],
                "summary": "List episodes",
                "parameters": [
                    {
                        "enum": [
                            "IDEA_BACKLOG",
                            "SCRIPT_DRAFT",
                            "SCRIPT_APPROVED",
                            "VO_GENERATED",
                            "MUSIC_GENERATED",
                            "ANIMATION_RENDERED",
                            "ASSEMBLED",
                            "QA_REVIEW",
                            "PLATFORM_ADAPTED",
                            "SCHEDULED",
                            "PUBLISHED",
                            "ANALYZED",
                            "CANCELLED"
                        ],
                        "type": "string",
                        "description": "Filter by pipeline status",
                        "name": "status",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/main.Episode"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Creates an episode in IDEA_BACKLOG. Status from the request body is ignored — every episode enters the pipeline at the first stage.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pipeline-episodes"
                ],
                "summary": "Create an episode",
                "parameters": [
                    {
                        "description": "Episode",
                        "name": "episode",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.Episode"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/main.Episode"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/pipeline/episodes/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pipeline-episodes"
                ],
                "summary": "Get one episode",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Episode ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.Episode"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Updates the editorial fields of an episode. Status is not editable here — use the transition endpoint.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pipeline-episodes"
                ],
                "summary": "Update episode content",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Episode ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Episode fields to update",
                        "name": "episode",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.Episode"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.Episode"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pipeline-episodes"
                ],
                "summary": "Delete an episode",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Episode ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Deleted"
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/pipeline/episodes/{id}/events": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pipeline-episodes"
                ],
                "summary": "Get the stage history of an episode",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Episode ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/main.EpisodeEvent"
                            }
                        }
                    }
                }
            }
        },
...
        "/pipeline/episodes/{id}/transition": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Used by production agents. Refuses any move the state machine does not allow, and refuses the two gate transitions, which require a human approval instead.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pipeline-episodes"
                ],
                "summary": "Move an episode to the next stage",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Episode ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Target stage",
                        "name": "transition",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.TransitionRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.Episode"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/pipeline/queue": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Read-only view of everything waiting at a stage, including which worker holds each episode. To actually take work use POST /pipeline/queue/claim, which hands one episode to one worker.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pipeline"
                ],
                "summary": "Look at the work queue for a stage",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Pipeline stage to poll",
                        "name": "status",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/main.Episode"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/pipeline/stages": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Returns every stage with its owning agent, allowed next stages and any human gate. Lets an agent discover where it fits without hardcoding the flow.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pipeline"
                ],
                "summary": "Describe the pipeline",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/main.StageInfo"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "main.Episode": {
            "type": "object",
            "properties": {
                "ai_disclosure": {
                    "type": "boolean"
                },
                "attempts": {
                    "description": "Attempts counts how many times the current stage has been tried. It\nresets whenever the episode moves on, so it measures this stage rather\nthan the episode's whole history.",
                    "type": "integer"
                },
                "claimed_at": {
                    "type": "string"
                },
                "claimed_by": {
                    "description": "ClaimedBy is the worker currently holding this episode, and ClaimedAt\nis when it took it. The claim is a lease: it expires so a worker that\ndies mid-job does not strand the episode. Both are cleared whenever the\nepisode changes stage.",
                    "type": "string"
                },
                "code": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "curriculum_topic_id": {
                    "type": "integer"
                },
                "duration_seconds": {
                    "type": "integer"
                },
                "failed_from": {
                    "$ref": "#/definitions/main.EpisodeStatus"
                },
                "id": {
                    "type": "integer"
                },
                "last_error": {
                    "type": "string"
                },
                "made_for_kids": {
                    "description": "MadeForKids drives the YouTube selfDeclaredMadeForKids flag. It defaults\nto true because every episode of this show is children's content, and\ngetting this wrong is a COPPA problem rather than a cosmetic one.",
                    "type": "boolean"
                },
                "published_at": {
                    "type": "string"
                },
                "scheduled_at": {
                    "type": "string"
                },
                "script_text": {
                    "type": "string"
                },
                "song_lyrics": {
                    "type": "string"
                },
                "song_title": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/main.EpisodeStatus"
                },
                "synopsis": {
                    "type": "string"
                },
                "tiktok_video_id": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "youtube_video_id": {
                    "type": "string"
                }
            }
        },
        "main.EpisodeEvent": {
            "type": "object",
            "properties": {
                "actor": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "episode_id": {
                    "type": "integer"
                },
                "from_status": {
                    "$ref": "#/definitions/main.EpisodeStatus"
                },
                "id": {
                    "type": "integer"
                },
                "note": {
                    "type": "string"
                },
                "to_status": {
                    "$ref": "#/definitions/main.EpisodeStatus"
                }
            }
        },
        "main.EpisodeStatus": {
            "type": "string",
            "enum": [
                "IDEA_BACKLOG",
                "SCRIPT_DRAFT",
                "SCRIPT_APPROVED",
                "VO_GENERATED",
                "MUSIC_GENERATED",
                "ANIMATION_RENDERED",
                "ASSEMBLED",
                "QA_REVIEW",
                "PLATFORM_ADAPTED",
                "SCHEDULED",
                "PUBLISHED",
                "ANALYZED",
                "CANCELLED",
                "FAILED"
            ],
            "x-enum-varnames": [
                "StatusIdeaBacklog",
                "StatusScriptDraft",
                "StatusScriptApproved",
                "StatusVOGenerated",
                "StatusMusicGenerated",
                "StatusAnimationRendered",
                "StatusAssembled",
                "StatusQAReview",
                "StatusPlatformAdapted",
                "StatusScheduled",
                "StatusPublished",
                "StatusAnalyzed",
                "StatusCancelled",
                "StatusFailed"
            ]
        },
        "main.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "main.StageInfo": {
            "type": "object",
            "properties": {
                "layer": {
                    "type": "string"
                },
                "next_statuses": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/main.EpisodeStatus"
                    }
                },
                "owner_agent": {
                    "type": "string"
                },
                "requires_gate": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/main.EpisodeStatus"
                }
            }
        },
        "main.TransitionRequest": {
            "type": "object",
            "properties": {
                "actor": {
                    "type": "string"
                },
                "note": {
                    "type": "string"
                },
                "to_status": {
                    "$ref": "#/definitions/main.EpisodeStatus"
                },
                "worker_id": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}